secure_storage: split aead_get.c into subfiles
Make one file per getter function to compile them only when they
are used.
The code is only moved and unchanged.
